fix tools and context fetching
