+++ packages/ai/src/contexts/classes/agent-rule.ts
