+++ packages/ai/src/contexts/classes/error.ts
