+++ packages/ai/src/contexts/classes/image.ts
