+++ packages/ai/src/tools/toolset.ts
@@ -21,6 +21,7 @@ import {
     WebSearchTool,
     WriteFileTool,
 } from './classes';
+import type { BaseTool } from './models/base';
 
 // Helper function to convert tool classes to ToolSet
 function createToolSet(toolClasses: Array<{ toolName: string; getAITool: () => any }>): ToolSet {
@@ -57,7 +58,7 @@ const allToolClasses = [...readOnlyToolClasses, ...editOnlyToolClasses];
 
 export const readOnlyToolset: ToolSet = createToolSet(readOnlyToolClasses);
 export const allToolset: ToolSet = createToolSet(allToolClasses);
-export const TOOLS_MAP = new Map(allToolClasses.map(toolClass => [toolClass.toolName, toolClass]));
+export const TOOLS_MAP: Map<string, typeof BaseTool> = new Map(allToolClasses.map(toolClass => [toolClass.toolName, toolClass]));
 
 export function getToolClassesFromType(chatType: ChatType) {
     return chatType === ChatType.ASK ? readOnlyToolClasses : allToolClasses
