+++ packages/ai/test/contexts/file-context.test.ts
@@ -1,6 +1,6 @@
 import { MessageContextType, type FileMessageContext, type HighlightMessageContext } from '@onlook/models';
 import { describe, expect, test } from 'bun:test';
-import { FileContext } from '../../src/contexts/classes/file-context';
+import { FileContext } from '../../src/contexts/classes/file';
 
 describe('FileContext', () => {
     const createMockFileContext = (overrides: Partial<FileMessageContext> = {}): FileMessageContext => ({
