+++ packages/ai/test/prompt/prompt.test.ts
@@ -1,6 +1,8 @@
 import { MessageContextType } from '@onlook/models';
 import { describe, expect, test } from 'bun:test';
 import path from 'path';
+import { FileContext } from '../../src/contexts/classes/file';
+import { HighlightContext } from '../../src/contexts/classes/highlight-context';
 import {
     type HydrateMessageOptions,
     getCreatePageSystemPrompt,
@@ -8,8 +10,6 @@ import {
     getSummaryPrompt,
     getSystemPrompt,
 } from '../../src/prompt/provider';
-import { FileContext } from '../../src/contexts/classes/file-context';
-import { HighlightContext } from '../../src/contexts/classes/highlight-context';
 
 const __dirname = import.meta.dir;
